refactor tests
fix isReadOnly
