+++ test/results_test.dart
@@ -19,14 +19,15 @@
 // ignore_for_file: unused_local_variable
 
 import 'dart:io';
-import 'package:test/test.dart' hide throws;
+import 'package:test/test.dart' hide test, throws;
 import '../lib/realm.dart';
 import 'test.dart';
 
 Future<void> main([List<String>? args]) async {
+  print(args);
   print("Current PID $pid");
 
-  setupTests();
+  setupTests(args);
 
   test('Results all should not return null', () {
     var config = Configuration([Car.schema]);
